Removed default info and added Photo Gallery
Removed most of the default information on the website and included a bootstrap photo gallery with default images.

Some parts including the slider needs to be updated.
The booking feature needs to be added, services page needs to created and a contact page, An admin panel to manage bookings should also be incorporated.